Integration test minor fix move stored event type to commons-test package
